sam-python-producer/msk-pyproducer/rest_api/requirements.txt
kafka-python
flask-lambda